# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/projects/{projectId}/workflows:
    post:
      operationId: createWorkflowRun
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: projectId
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRunRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Create a workflow run and seed its initial steps
      tags:
      - workflows
...
  /api/v1/workflows:
    get:
      operationId: listWorkflowRuns
      parameters:
      - description: Project id filter.
        in: query
        name: projectId
        schema:
          description: Project id filter.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkflowsResponse'
          description: OK
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List workflow run summaries
      tags:
      - workflows
  /api/v1/workflows/{workflowId}:
    get:
      operationId: getWorkflowRun
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Get a workflow run with its steps and attempts
      tags:
      - workflows
  /api/v1/workflows/{workflowId}/cancel:
    post:
      operationId: cancelWorkflowRun
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Cancel a workflow run and its non-terminal steps
      tags:
      - workflows
  /api/v1/workflows/{workflowId}/continue:
    post:
      operationId: continueWorkflowRun
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: 'Continue a workflow run: dispatch its review step''s real Claude reviewer
        once the work step has completed (Checkpoint 8C). Idempotent no-op when nothing
        is currently dispatchable.'
      tags:
      - workflows
  /api/v1/workflows/{workflowId}/plan:
    get:
      operationId: getWorkflowPlan
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
      summary: Get a durable master plan and planned tasks
      tags:
      - workflows
  /api/v1/workflows/{workflowId}/plan/approve:
    post:
      operationId: approveWorkflowPlan
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
          description: OK
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
      summary: Approve a validated plan and dispatch the first eligible task
      tags:
      - workflows
  /api/v1/workflows/{workflowId}/plan/generate:
    post:
      operationId: generateWorkflowPlan
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
          description: OK
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
      summary: Generate and deterministically validate a durable master plan
      tags:
      - workflows
  /api/v1/workflows/{workflowId}/plan/reject:
    post:
      operationId: rejectWorkflowPlan
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
          description: OK
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
      summary: Reject and cancel a master plan
      tags:
      - workflows
...
  /api/v1/workflows/{workflowId}/start:
    post:
      operationId: startWorkflowRun
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRunResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: 'Start a pending workflow run: run its plan step and dispatch its work
        step''s Codex worker'
      tags:
      - workflows
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    CreateWorkflowRunRequest:
      properties:
        masterPlan:
          description: Generate a provider-neutral master plan before execution.
          type: boolean
        objective:
          description: The workflow run's objective.
          type: string
        planApprovalMode:
          enum:
          - manual
          - auto
          type: string
        verification:
          $ref: '#/components/schemas/WorkflowVerificationPlan'
      required:
      - objective
      type: object
...
    ListWorkflowsResponse:
      properties:
        workflows:
          items:
            $ref: '#/components/schemas/WorkflowRunView'
          type: array
      required:
      - workflows
      type: object
...
    WorkflowRunResponse:
      properties:
        workflow:
          $ref: '#/components/schemas/WorkflowRunDetailView'
      required:
      - workflow
      type: object
...
    WorkflowVerificationCommand:
      properties:
        args:
          items:
            type: string
          type: array
        command:
          type: string
        requiredExitCode:
          type: integer
        retrySafe:
          type: boolean
        timeoutSeconds:
          type: integer
        workingDirectory:
          type: string
      required:
      - command
      - requiredExitCode
      - retrySafe
      type: object
...
    WorkflowVerificationPlan:
      properties:
        commands:
          items:
            $ref: '#/components/schemas/WorkflowVerificationCommand'
          type: array
        files:
          items:
            $ref: '#/components/schemas/WorkflowVerificationFile'
          type: array
      type: object
...